Add UserVoice widget to global footer.
The widget is placed on the right hand side of the window, half-way down
the vertical space.  It is "fixed" to the window, so it will remain in
place as the user scrolls the page.

The widget might be better placed at the very bottom of the <body> tag in
_Layout.cshtml to improve page rendering.