Fix an error running on .NET 4.5.
The list of Members started failing when run on .NET 4.5, because the
PagedList helper library was causing a VerificationException to be thrown.
This new Exception type is thrown when the code being run is required to
be type safe, but cannot be verified as such.  See
http://msdn.microsoft.com/en-us/library/system.security.verificationexception.aspx
for more information.

Simply upgrading the PagedList helper library to the latest version
resolved the issue.